Add files via upload
As a hands on Assignment, This project allows to work on real world dataset provided by the Chicago Data Portal to improve educational outcomes for children and youth in the City of Chicago. Your job is to analyze the census, crime, and school data for a given neighborhood or district. It  identifies causes that impact the enrollment, safety, health, environment ratings of schools.